Add passwords directory
This is where password files are stored, but I've ignored them
all so they aren't added to the repo. I use the password files for
gpg2 to read and pass to programs. Example: mbsync can acquire a
password by using a command, rather than hardcoding the password
in the configuration file.

Future concept: Have a script located here that pulls the password
files from a secure source?