Update terminal open call from i3
The upgrade to Fedora 25 broke this on my desktop, which has both
gnome-terminal and urxvt. Because I don't yet have a proper
configuration set up for urxvt, it's now hardcoded to gnome-terminal.

My go-to install of Linux is Fedora, especially where I use i3, so this
should not affect anything else that I run.